Simplify ros config
The enter_env script was doing too much work. Deciding on whether or not we should launch a roscore server in a generic way is an impossible problem. Each problem has its own requirements. 

Thew new system is a simple `enter_env` that just sources (automatically) the appropriate setup files and `start_roscore` script to start  a `roscore` in the background.