install: Create required directories ahead of time
There are two sorts of directories that need to exist:

- directories that we expect to exist (mostly XDG directories:
~/.config/, ~/.local/bin/, ...)
- directories that need to be created for a tool

It used to be that the format were created by tools (_e.g._ installing
ish creates .local/bin, but that is not very intuitive behviour. To fix
this, we create `REQUIRED_DIRS` that are always created whatever the
profile we install is.

In a second phase, we create the dirs reqiured by the individual tools,
but we create them before actually installing anything. The rationale
here is week, but can be helpful if we add depdencies between tools to
ensure that all the preconditions are met before installing a tool.
Until then, I struggle to justify the change, but it's here now ...